+++ spec/unit/kebabcase_spec.rb
@@ -23,7 +23,7 @@ RSpec.describe Strings::Case, "#kebabcase" do
     "get2HTTPResponse" => "get2-http-response",
     "HTTPResponseCode" => "http-response-code",
     "HTTPResponseCodeXY" => "http-response-code-xy",
-    "supports IPv6 on iOS?" => "supports-ipv6-on-ios"
+    "supports IPv6 on iOS?" => "supports-i-pv6-on-i-os"
   }.each do |actual, expected|
     it "applies kebabcase from #{actual.inspect} to #{expected.inspect}" do
       expect(Strings::Case.kebabcase(actual)).to eq(expected)